Updated Sidebar features
All options change color on hover. Have temporary color change for bm-items, dropdown-options and dbutton to better contrast them against background (Must be changed to better color), added an arrow next to Conversions to indicate it is a dropdown menu that needs to be clicked (dbutton:before), and added .show, which is used to show/hide dropdown menu options.